+++ azext_iot/sdk/iothub/mgmt/_configuration.py
@@ -28,13 +28,13 @@ class IotHubClientConfiguration:  # pylint: disable=too-many-instance-attributes
     :type credential: ~azure.core.credentials.TokenCredential
     :param subscription_id: The subscription identifier. Required.
     :type subscription_id: str
-    :keyword api_version: Api Version. Default value is "2025-08-01-preview". Note that overriding
+    :keyword api_version: Api Version. Default value is "2026-03-01-preview". Note that overriding
      this default value may result in unsupported behavior.
     :paramtype api_version: str
     """
 
     def __init__(self, credential: "TokenCredential", subscription_id: str, **kwargs: Any) -> None:
-        api_version: str = kwargs.pop("api_version", "2025-08-01-preview")
+        api_version: str = kwargs.pop("api_version", "2026-03-01-preview")
 
         if credential is None:
             raise ValueError("Parameter 'credential' must not be None.")